+++ crates/cargo-contract/src/cmd/extrinsics/instantiate.rs
@@ -17,7 +17,6 @@
 use super::{
     display_contract_exec_result,
     prompt_confirm_tx,
-    runtime_api::api,
     state_call,
     submit_extrinsic,
     Balance,
@@ -32,12 +31,15 @@ use super::{
     MAX_KEY_COL_WIDTH,
 };
 use crate::{
-    cmd::extrinsics::{
-        display_contract_exec_result_debug,
-        display_dry_run_result_warning,
-        events::DisplayEvents,
-        ErrorVariant,
-        TokenMetadata,
+    cmd::{
+        extrinsics::{
+            display_contract_exec_result_debug,
+            display_dry_run_result_warning,
+            events::DisplayEvents,
+            ErrorVariant,
+            TokenMetadata,
+        },
+        runtime_api::api,
     },
     DEFAULT_KEY_COL_WIDTH,
 };