+++ crates/cargo-contract/src/cmd/extrinsics/upload.rs
@@ -16,10 +16,6 @@
 
 use super::{
     display_dry_run_result_warning,
-    runtime_api::api::{
-        self,
-        runtime_types::pallet_contracts::wasm::Determinism,
-    },
     state_call,
     submit_extrinsic,
     Balance,
@@ -31,10 +27,16 @@ use super::{
     TokenMetadata,
 };
 use crate::{
-    cmd::extrinsics::{
-        events::DisplayEvents,
-        ErrorVariant,
-        WasmCode,
+    cmd::{
+        extrinsics::{
+            events::DisplayEvents,
+            ErrorVariant,
+            WasmCode,
+        },
+        runtime_api::api::{
+            self,
+            runtime_types::pallet_contracts::wasm::Determinism,
+        },
     },
     name_value_println,
 };
@@ -158,7 +160,7 @@ impl UploadCommand {
         let token_metadata = TokenMetadata::query(client).await?;
         let storage_deposit_limit =
             self.extrinsic_opts.storage_deposit_limit(&token_metadata)?;
-        let call = super::runtime_api::api::tx().contracts().upload_code(
+        let call = crate::cmd::runtime_api::api::tx().contracts().upload_code(
             code.0,
             storage_deposit_limit,
             Determinism::Deterministic,